Extract editing methods from input_field/mod.rs
Move 9 cursor movement and text editing helper methods from
InputFieldState in mod.rs (1000 lines) to a new editing.rs submodule
(139 lines), bringing mod.rs down to 870 lines.

Follows the same pattern as text_area/cursor.rs — extracted methods use
pub(super) visibility so they remain accessible from mod.rs where
update() calls them.
